# src/main/resources/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ousmane Vincent Dione",
    "description": "OpenApi documentation for Stock Management",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ousmane Vincent",
      "url": "https://ndiolfouta.com",
      "email": "dioneousmanevincent@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local ENV"
    },
    {
      "url": "https://ndiolfouta.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/ventes": {
      "post": {
        "tags": [
          "vente-api"
        ],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/utilisateurs": {
      "post": {
        "tags": [
          "utilisateur-api"
        ],
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ventes/{idVente}": {
      "get": {
        "tags": [
          "vente-api"
        ],
        "operationId": "getVenteById",
        "parameters": [
          {
            "name": "idVente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/utilisateurs/{idUtilisateur}": {
      "get": {
        "tags": [
          "utilisateur-api"
        ],
        "operationId": "getUtilisateurById",
        "parameters": [
          {
            "name": "idUtilisateur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/ventes/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueVente",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneVenteDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/commandeFournisseur/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueCommandeFournisseur",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeFournisseurDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/commandeClients/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueCommandeClient",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeClientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/filter/category/{idCategory}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueByIdCatergory",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ventes/delete/{idVente}": {
      "delete": {
        "tags": [
          "vente-api"
        ],
        "operationId": "deleteVente",
        "parameters": [
          {
            "name": "idVente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/utilisateurs/delete/{idUtilisateur}": {
      "delete": {
        "tags": [
          "utilisateur-api"
        ],
        "operationId": "deleteUtilisateur",
        "parameters": [
          {
            "name": "idUtilisateur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commandeClients/deleteArticle/{idCommande}/{idLigneCommande}": {
      "delete": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateVente": {
            "type": "string",
            "format": "date-time"
          },
          "commentaire": {
            "type": "string"
          }
        }
      },
      "AdresseDto": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "EntrepriseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "codeFiscal": {
            "type": "string"
          },
          "siteWeb": {
            "type": "string"
          },
          "adresseDto": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "utilisateurDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UtilisateurDto"
            }
          }
        }
      },
      "UtilisateurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "entrepriseDto": {
            "$ref": "#/components/schemas/EntrepriseDto"
          }
        }
      },
      "CommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "fournisseurDto": {
            "$ref": "#/components/schemas/FournisseurDto"
          }
        }
      },
      "FournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresseDto": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "commandeFournisseurDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandeFournisseurDto"
            }
          }
        }
      },
      "Adresse": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/Adresse"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          }
        }
      },
      "CommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "clientDto": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "etat": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          }
        }
      },
...
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codeArticle": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "prixUnitaireHt": {
            "type": "number"
          },
          "tauxTval": {
            "type": "number"
          },
          "prixUnitaireTtc": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "categoryDto": {
            "$ref": "#/components/schemas/CategoryDto"
          }
        }
      },
      "LigneCommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleDto": {
            "$ref": "#/components/schemas/ArticleDto"
          },
          "commandeClientDto": {
            "$ref": "#/components/schemas/CommandeClientDto"
          }
        }
      },
      "LigneVenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "venteDto": {
            "$ref": "#/components/schemas/VenteDto"
          },
          "quantite": {
            "type": "number"
          },
          "prixUnitaire": {
            "type": "number"
          }
        }
      },
      "LigneCommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleDto": {
            "$ref": "#/components/schemas/ArticleDto"
          },
          "commandeFournisseurDto": {
            "$ref": "#/components/schemas/CommandeFournisseurDto"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Gestion stock auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}